split out static options into a separate file
